Initial Revision
Defines a  set of rdf options that can be applied together to methods, messages, and fields to annotate their properties.

The next step will be development of a set of example services that use the option definitions to expose semantic information

Finally, a tool will be implemented to extract the semantic information from generated coded and render that information into triples.